Move to the config package.
