+++ app/org/sagebionetworks/bridge/config/ConfigReader.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.bridge.context;
+package org.sagebionetworks.bridge.config;
 
 public interface ConfigReader {
 	public String read(String name);
